用户改为管理员

+++ core-service-api/src/main/java/com/jdkcc/ts/service/dto/request/wxuser/WxUserSearchReq.java
@@ -20,7 +20,7 @@ public class WxUserSearchReq implements java.io.Serializable {
     /**
      * 性别
      */
-	private Short sex;
+	private Integer sex;
     /**
      * 国家
      */
