+++ core-service-api/src/main/java/com/jdkcc/ts/service/dto/request/wxuser/WxUserUpdateReq.java
@@ -19,7 +19,7 @@ public class WxUserUpdateReq implements java.io.Serializable {
     /**
      * 性别
      */
-	private Short sex;
+	private Integer sex;
     /**
      * 国家
      */
