+++ core-service-api/src/main/java/com/jdkcc/ts/service/dto/response/WxUserDTO.java
@@ -25,7 +25,7 @@ public class WxUserDTO implements java.io.Serializable {
     /**
      * 性别
      */
-	private Short sex;
+	private Integer sex;
     /**
      * 国家
      */
